+++ content-author.php
@@ -1,8 +1,3 @@
-<?php
-/**
- * @version 6.4.3
- */
-?>
 			<?php if ( is_singular() && get_the_author_meta( 'description' ) && is_multi_author() ) : ?>
 				<div class="author-info small-12 large-12 cell">
 					<div class="author-avatar small-12 medium-12 large-12 columns">