+++ content-image.php
@@ -1,8 +1,3 @@
-<?php
-/**
- * @version 6.4.3
- */
-?>
 	<article id="post-<?php the_ID(); ?>" <?php post_class(); ?> <?php wpforge_article_schema( 'CreativeWork' ); ?>>
 		<div class="entry-meta-header">
 			<?php  if( get_theme_mod( 'wpforge_cat_display','yes' ) == 'yes') { ?>