+++ content-none.php
@@ -1,8 +1,3 @@
-<?php
-/**
- * @version 6.4.3
- */
-?>
 	<article id="post-0" class="post no-results not-found">
 		<?php if ( current_user_can( 'edit_posts' ) ) : ?>
 			<header class="entry-header">