+++ page-templates/front-page.php
@@ -1,7 +1,6 @@
 <?php
 /**
  * Template Name: Front Page Template
- * @version 6.4.3
  */
 get_header(); ?>
 	<div id="content" class="small-12 large-12 cell" role="main">