+++ searchform.php
@@ -1,8 +1,3 @@
-<?php
-/**
- * @version 6.4.3
- */
-?>
 <form role="search" method="get" id="searchform" action="<?php echo home_url('/'); ?>">
 	<div class="grid-container">
 	<div class="grid-x grid-margin-x large-margin-collapse">