+++ sidebar-footer.php
@@ -1,7 +1,4 @@
 <?php
-/**
- * @version 6.4.3
- */
 if ( ! is_active_sidebar( 'footer-sidebar-1' )
 	&& ! is_active_sidebar( 'footer-sidebar-2' )
 	&& ! is_active_sidebar( 'footer-sidebar-3' )